Started rewrite of tsp
git-svn-id: http://zhihaolou.org/svn/neoParSA/trunk@222 a3e5a195-83b0-4c6b-8637-e6d081586ea5

+++ parsa/simPLam.cpp
@@ -77,7 +77,7 @@ double simPLam::updateS(const aState& state)
 
 void simPLam::calcStats(int nsteps)
 {
-    double from[3] = { sum, sumsq, success }, to[3];
+    double from[3] = { sum, sumsq, (double)success }, to[3];
     double N = (double) (mpiState.nnodes) * nsteps;
     MPI_Allreduce(from, to, 3, MPI_DOUBLE, MPI_SUM, mpiState.comm);
     mean = to[0] / N;
